Общие методы перенесены в BaseTest
