+++ src/test/java/autotests/clients/ActionsClient.java
@@ -1,5 +1,6 @@
 package autotests.clients;
 
+import autotests.BaseTest;
 import autotests.EndpointConfig;
 import com.consol.citrus.TestCaseRunner;
 import com.consol.citrus.http.client.HttpClient;
@@ -23,119 +24,7 @@ import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 
 @ContextConfiguration(classes = {EndpointConfig.class})
 @Epic("Тесты на duck-action-controller")
-public class ActionsClient extends TestNGCitrusSpringSupport {
-    @Autowired
-    protected HttpClient duckService;
-
-    @Autowired
-    protected SingleConnectionDataSource testDB;
-
-
-    //Создание утки в БД
-    @Step("SQL запрос создания уточки в БД")
-    public void createDuckDB(TestCaseRunner runner, String id,
-                             String color, double height,
-                             String material, String sound,
-                             String wings_state) {
-        runner.$(sql(testDB)
-                .statement("INSERT INTO duck "
-                        + "(id, color, height, material, sound, wings_state)\n"
-                        + "VALUES ('" + id + "', '" + color + "', '" + height + "', '"
-                        + material + "', '" + sound + "', '" + wings_state + "');"));
-    }
-
-    //Удаление утки из БД
-    @Step("SQL запрос удаления уточки из БД")
-    public void deleteDuckDB(TestCaseRunner runner, String id) {
-        runner.$(sql(testDB)
-                .statement("DELETE FROM duck WHERE ID=" + id));
-    }
-
-    //Создание утки через API
-    @Step("POST запрос на создание уточки")
-    public void createDuckAPI(TestCaseRunner runner, Object body) {
-        runner.$(http().client(duckService)
-                .send()
-                .post("/api/duck/create")
-                .message()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(new ObjectMappingPayloadBuilder(body, new ObjectMapper())));
-    }
-
-    //Генерация случайного числа, чётного (evenFlag = true) или нечётного (evenFlag = false)
-    @Step("Генерация чётного/нечётного числа")
-    public int generateEvenOddNumber(boolean evenFlag) {
-        int randomValue;
-        int minValue = 1;
-        int maxValue = 1000000000;
-
-        do {
-            randomValue = minValue + (int) (Math.random() * (maxValue - minValue + 1));
-        }
-        while ((randomValue % 2 != 0) == evenFlag);
-        return randomValue;
-    }
-
-    //Создание тестовой переменной ID уточки
-    @Step("Запись ID уточки из ответа в тестовую переменную")
-    public void createTestVariableDuckId(TestCaseRunner runner) {
-        runner.$(http().client(duckService)
-                .receive()
-                .response(HttpStatus.OK)
-                .message()
-                .extract(fromBody().expression("$.id", "duckId")));
-    }
-
-    //Преобразование контекстной переменной в тип long
-    @Step("Преобразование контекстной переменной в тип long")
-    public long getLongTestVariable(TestCaseRunner runner, String testVariable) {
-        AtomicLong id = new AtomicLong();
-        runner.$(action -> {
-            id.set(Long.parseLong(action.getVariable(testVariable)));
-        });
-        return id.longValue();
-    }
-
-    //Инкремент контекстной переменной
-    @Step("Инкремент контекстной переменной")
-    public void setIncrementTestVariable(TestCaseRunner runner, String testVariable) {
-        runner.$(action -> {
-            action.setVariable(testVariable, getLongTestVariable(runner, testVariable) + 1);
-        });
-    }
-
-    //Валидация ответа с передачей ответа String’ой
-    @Step("Валидация ответа с передачей ожидаемого ответа String’ой")
-    public void validateResponseString(TestCaseRunner runner, String responseMessage) {
-        runner.$(http().client(duckService)
-                .receive()
-                .response(HttpStatus.OK)
-                .message()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(responseMessage));
-    }
-
-    //Валидация ответа с передачей ответа из папки Payload
-    @Step("Валидация ответа с передачей ожидаемого ответа из папки Payload")
-    public void validateResponsePayload(TestCaseRunner runner, Object body) {
-        runner.$(http().client(duckService)
-                .receive()
-                .response(HttpStatus.OK)
-                .message()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(new ObjectMappingPayloadBuilder(body, new ObjectMapper())));
-    }
-
-    //Валидация ответа с передачей ответа из папки Resources
-    @Step("Валидация ответа с передачей ожидаемого ответа из папки Resources")
-    public void validateResponseResources(TestCaseRunner runner, String responseMessage) {
-        runner.$(http().client(duckService)
-                .receive()
-                .response(HttpStatus.OK)
-                .message()
-                .contentType(MediaType.APPLICATION_JSON_VALUE) //spring
-                .body(new ClassPathResource(responseMessage)));
-    }
+public class ActionsClient extends BaseTest {
 
 
     //Формирование звука
